unary cleanup
