+++ tests/tt_metal/tt_metal/test_kernels/compute/dropout_sfpu.cpp
@@ -17,7 +17,7 @@ void MAIN {
     uint32_t int_scale_factor = get_compile_time_arg_val(4);
 
     init_sfpu(tt::CBIndex::c_0, tt::CBIndex::c_16);
-    dropout_tile_init(seed);
+    dropout_kernel_init(seed);
     for (uint32_t block_index = 0; block_index < per_core_block_cnt; block_index++) {
         cb_reserve_back(tt::CBIndex::c_16, per_core_block_dim);
         for (uint32_t tile_index = 0; tile_index < per_core_block_dim; ++tile_index) {
