+++ tests/ttnn/unit_tests/operations/test_dropout.py
@@ -28,7 +28,7 @@ def test_dopout(device):
     s = 124
     prob = 0.2
     for _ in range(1000):
-        output = ttnn.dropout(t_tt, seed=s, probability=prob, scale=1.0 / (1.0 - prob))
+        output = ttnn.experimental.dropout(t_tt, seed=s, probability=prob, scale=1.0 / (1.0 - prob))
         output_torch = ttnn.to_torch(output)
         r = 1.0 - (torch.count_nonzero(output_torch) / torch.count_nonzero(t)).item()
         tt_ratios.append(r)
