+++ tt-train/sources/ttml/core/ttnn_all_includes.hpp
@@ -41,6 +41,7 @@
 #include <ttnn/operations/eltwise/unary_backward/unary_backward.hpp>                               // NOLINT
 #include <ttnn/operations/embedding/embedding.hpp>                                                 // NOLINT
 #include <ttnn/operations/embedding_backward/embedding_backward.hpp>                               // NOLINT
+#include <ttnn/operations/experimental/dropout/dropout.hpp>                                        // NOLINT
 #include <ttnn/operations/experimental/transformer/nlp_concat_heads/nlp_concat_heads.hpp>          // NOLINT
 #include <ttnn/operations/experimental/transformer/nlp_create_qkv_heads/nlp_create_qkv_heads.hpp>  // NOLINT
 #include <ttnn/operations/full_like/full_like.hpp>                                                 // NOLINT
