+++ tt-train/tests/ttnn_fixed/dropout_op_test.cpp
@@ -2,7 +2,6 @@
 
 #include <core/ttnn_all_includes.hpp>
 #include <ttnn/operations/core/compute_kernel/compute_kernel_config.hpp>
-#include <ttnn/operations/experimental/dropout/dropout_op.hpp>
 #include <ttnn/operations/reduction/generic/generic_reductions.hpp>
 
 #include "autograd/auto_context.hpp"
