+++ ttnn/cpp/ttnn/operations/eltwise/unary/unary.cpp
@@ -387,39 +387,6 @@ Tensor Ceil::invoke(
         DefaultQueueId, input_tensor, {UnaryWithParam{op_type}}, memory_config, optional_output_tensor);
 }
 
-Tensor Dropout::invoke(
-    const Tensor& input,
-    const uint32_t seed,
-    const float probability,
-    const float scale,
-    const std::optional<MemoryConfig>& memory_config,
-    const std::optional<Tensor>& optional_output_tensor) {
-    TT_ASSERT(input.device()->arch() != tt::ARCH::GRAYSKULL, "Dropout is not currently supported on Grayskull");
-    return detail::unary_impl(
-        DefaultQueueId,
-        input,
-        {UnaryWithParam{UnaryOpType::DROPOUT, {static_cast<float>(seed), probability, scale}}},
-        memory_config,
-        optional_output_tensor);
-}
-
-Tensor Dropout::invoke(
-    uint8_t queue_id,
-    const Tensor& input,
-    const uint32_t seed,
-    const float probability,
-    const float scale,
-    const std::optional<MemoryConfig>& memory_config,
-    const std::optional<Tensor>& optional_output_tensor) {
-    TT_ASSERT(input.device()->arch() != tt::ARCH::GRAYSKULL, "Dropout is not currently supported on Grayskull");
-    return detail::unary_impl(
-        queue_id,
-        input,
-        {UnaryWithParam{UnaryOpType::DROPOUT, {static_cast<float>(seed), probability, scale}}},
-        memory_config,
-        optional_output_tensor);
-}
-
 template <UnaryOpType unary_op_type, typename T>
 Tensor ExecuteUnaryWithIntegerParameter<unary_op_type, T>::invoke(
     uint8_t queue_id,
