+++ ttnn/cpp/ttnn/operations/experimental/dropout/device/kernels/compute/dropout_kernel.cpp
@@ -19,7 +19,7 @@ void MAIN {
     uint32_t seed = get_arg_val<uint32_t>(0);
 
     init_sfpu(tt::CBIndex::c_0, tt::CBIndex::c_2);
-    dropout_tile_init(seed);
+    dropout_kernel_init(seed);
     for (uint32_t block_index = 0; block_index < per_core_block_cnt; block_index++) {
         cb_reserve_back(tt::CBIndex::c_2, per_core_block_dim);
         for (uint32_t tile_index = 0; tile_index < per_core_block_dim; ++tile_index) {
