+++ ttnn/cpp/ttnn/operations/experimental/dropout/dropout.hpp
@@ -1,0 +1,17 @@
+
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+namespace ttnn::operations::experimental {
+
+struct DropoutOperation {
+    static Tensor invoke(const Tensor& input_tensor, float prob, float scale, uint32_t seed);
+};
+}  // namespace ttnn::operations::experimental
+namespace ttnn::experimental {
+constexpr auto dropout =
+    ttnn::register_operation<"ttnn::::experimental::dropout", ttnn::operations::experimental::DropoutOperation>();
+}  // namespace ttnn::experimental
