+++ ttnn/cpp/ttnn/operations/experimental/dropout/dropout_pybind.hpp
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+#include "pybind11/pybind_fwd.hpp"
+
+namespace ttnn::operations::experimental::dropout::detail {
+namespace py = pybind11;
+void bind_experimental_dropout_operation(py::module& module);
+
+}  // namespace ttnn::operations::experimental::dropout::detail
